+++ lib/constants/constants.dart
@@ -7,7 +7,7 @@ const double textSizeMax = 60.0;
 const int appShortcutMin = 0;
 const int appShortcutMax = 10;
 
-const int pixelationLevelMin = 1;
-const int pixelationLevelMax = 100;
+const int pixelationLevelMin = 5;
+const int pixelationLevelMax = 50;
 
 const String customAppNames = 'custom_app_names';